+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
-* Fix rustsec adivisory warning by replacing `paste` dependency with the new maintained `pastey` crate. 
+* Add functions `all`, `all_ok`, `all_some`, `any`, `any_ok`, `any_some` in `zng::task`. These functions are dynamic versions of the
+* Task macros `all!`, `all_ok!`, `all_some!`, `any!`, `any_ok!` and `any_some` now accept `IntoFuture` inputs.
 
 # 0.14.0